Release 0.20.0 (#298)
* feat(prompts): support `system message` in `MoxinInstructPrompt`


* chore(core): fix clippy issue


* chore(core): remove `</think>` from the generation of `qwen3`


* refactor!(endpoints): rename `name` field of `IndexRequest` to `index`


* doc(prompts): update `README`


* chore(prompts): update `MoxinInstructPrompt`


* version(prompts): bump to `0.30.2`


* version(endpoints): bump to `0.28.0`


* version(core): bump to `0.32.8`


* chore(workspace): upgrade `endpoints` dep


* version: bump to `0.20.0`


---------

+++ crates/llama-core/src/rag.rs
@@ -239,7 +239,7 @@ pub async fn rag_retrieve_context(
                     if let Some(source) = payload.get("source").and_then(Value::as_str) {
                         points.push(RagScoredPoint {
                             source: source.to_string(),
-                            score: point.score,
+                            score: point.score as f64,
                         })
                     }
